Updates to striplog detection methods; Code refactoring
