+++ src/extraction/main.py
@@ -6,14 +6,13 @@ import os
 from pathlib import Path
 
 import click
-import cv2
 import pymupdf
 from dotenv import load_dotenv
 from tqdm import tqdm
 
 from extraction import DATAPATH
 from extraction.annotations.draw import draw_predictions, plot_strip_logs, plot_tables
-from extraction.annotations.plot_utils import plot_lines
+from extraction.annotations.plot_utils import plot_lines, save_visualization
 from extraction.evaluation.benchmark.score import evaluate_all_predictions
 from extraction.features.extract import extract_page
 from extraction.features.groundwater.groundwater_extraction import (
@@ -26,10 +25,9 @@ from extraction.features.predictions.file_predictions import FilePredictions
 from extraction.features.predictions.overall_file_predictions import OverallFilePredictions
 from extraction.features.predictions.predictions import BoreholeListBuilder
 from extraction.features.stratigraphy.layer.layer import LayersInDocument
-from extraction.features.utils.geometry.circle_detection import extract_circles
 from extraction.features.utils.geometry.line_detection import extract_lines
+from extraction.features.utils.strip_log_detection import detect_strip_logs
 from extraction.features.utils.table_detection import (
-    detect_strip_logs,
     detect_structure_lines,
     detect_table_structures,
 )
@@ -323,9 +321,7 @@ def start_pipeline(
                 table_structures = detect_table_structures(page_index, doc, structure_lines, text_lines)
 
                 # Detect strip logs on the page
-                geometric_circles = extract_circles(page, line_detection_params, text_lines)
-                # strip_logs = detect_strip_logs(page_index, doc, structure_lines, geometric_circles, text_lines)
-                strip_logs = detect_strip_logs(structure_lines, geometric_circles, text_lines)
+                strip_logs = detect_strip_logs(page, geometric_lines, line_detection_params, text_lines)
 
                 # extract the statigraphy
                 page_layers = extract_page(
@@ -353,35 +349,45 @@ def start_pipeline(
                 )
                 all_groundwater_entries.groundwater_feature_list.extend(groundwater_entries)
 
+                # # Draw table structures if requested
+                # if draw_tables:
+                #     img = plot_tables(page, table_structures, page_index)
+
+                #     if draw_directory:
+                #         table_img_path = draw_directory / f"{Path(filename).stem}_page_{page.number + 1}_tables.png"
+                #         cv2.imwrite(str(table_img_path), img)
+
+                #     if mlflow_tracking:
+                #         mlflow.log_image(img, f"pages/{filename}_page_{page.number + 1}_tables.png")
+
+                #     elif not draw_directory:
+                #         logger.warning("draw_directory is not defined. Skipping saving table image.")
+
+                # # Draw strip logs if requested
+                # if draw_strip_logs:
+                #     img = plot_strip_logs(page, strip_logs, page_index)
+
+                #     if draw_directory:
+                #         strip_img_path = (
+                #             draw_directory / f"{Path(filename).stem}_page_{page.number + 1}_strip_logs.png"
+                #         )
+                #         cv2.imwrite(str(strip_img_path), img)
+
+                #     if mlflow_tracking:
+                #         mlflow.log_image(img, f"pages/{filename}_page_{page.number + 1}_strip_logs.png")
+
+                #     elif not draw_directory:
+                #         logger.warning("draw_directory is not defined. Skipping saving table image.")
+
                 # Draw table structures if requested
                 if draw_tables:
                     img = plot_tables(page, table_structures, page_index)
-
-                    if draw_directory:
-                        table_img_path = draw_directory / f"{Path(filename).stem}_page_{page.number + 1}_tables.png"
-                        cv2.imwrite(str(table_img_path), img)
-
-                    if mlflow_tracking:
-                        mlflow.log_image(img, f"pages/{filename}_page_{page.number + 1}_tables.png")
-
-                    elif not draw_directory:
-                        logger.warning("draw_directory is not defined. Skipping saving table image.")
+                    save_visualization(img, filename, page.number + 1, "tables", draw_directory, mlflow_tracking)
 
                 # Draw strip logs if requested
                 if draw_strip_logs:
                     img = plot_strip_logs(page, strip_logs, page_index)
-
-                    if draw_directory:
-                        strip_img_path = (
-                            draw_directory / f"{Path(filename).stem}_page_{page.number + 1}_strip_logs.png"
-                        )
-                        cv2.imwrite(str(strip_img_path), img)
-
-                    if mlflow_tracking:
-                        mlflow.log_image(img, f"pages/{filename}_page_{page.number + 1}_strip_logs.png")
-
-                    elif not draw_directory:
-                        logger.warning("draw_directory is not defined. Skipping saving table image.")
+                    save_visualization(img, filename, page.number + 1, "strip_logs", draw_directory, mlflow_tracking)
 
                 if draw_lines:  # could be changed to if draw_lines and mlflow_tracking:
                     if not mlflow_tracking:
